Rename from pbinfo to Svcdef
