+++ svcdef/test-go.txt
